backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-ollama
chromadb
pypdf
python-multipart
pillow
langgraph
tavily-python
openai-whisper
soundfile
duckduckgo-search
mlx-lm
datasets
huggingface_hub
langchain-chroma
langchain-ollama
langchain-huggingface
langchain-google-genai
